requirements.txt
torch==2.10.0
torchvision==0.25.0
torchaudio==2.10.0
torchxrayvision==1.4.0
captum==0.8.0
git+https://github.com/jacobgil/pytorch-grad-cam.git
scikit-learn==1.8.0
pandas==3.0.1
numpy==1.26.4
matplotlib==3.10.8
tqdm==4.67.3
Pillow==12.0.0
opencv-python==4.11.0.86
pycocotools==2.0.11
rich==14.3.3
